DoublyLinkedList
Here I implemented the doubly linked list.  (perfect for tracking the order of elements, and changing the order, in constant time), (swaps can be done in O(1)) 
It can be used together with HashMap, where we have the keys of some type, and values of type doublyLinkedList, that way, we can 
keep track of the order, all done in O(1) time, as we can access nodes in constant time via HashMap, and then perform either delete 
or say, insertBefore functions, which again are O(1).